Init paypal on beforeRegistration hook.
